# src/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Numgas API Contract
  version: 1.0.0
  description: >-
    Kontrak API untuk sistem Numgas meliputi Auth & Account, Public Access,
    SPPG Dashboard, dan School Dashboard.
servers:
  - url: /api/v1
    description: Default API base path
...
paths:
...
  /menu/upload:
    post:
      tags:
        - 03 - SPPG Dashboard
      summary: Upload Menu CSV
      operationId: uploadMenuCsv
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File CSV menu yang mengikuti template SIMBA.
              required:
                - file
      responses:
        '201':
          description: CSV menu berhasil diunggah.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /school/dashboard/summary:
    get:
      tags:
        - 04 - School Dashboard
      summary: School Summary
      operationId: getSchoolDashboardSummary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ringkasan dashboard sekolah berhasil diambil.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SchoolSummary'
                required:
                  - data
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Permintaan tidak valid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Tidak terautentikasi atau token tidak valid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    SchoolSummary:
      type: object
      properties:
        schoolName:
          type: string
        address:
          type: string
        partnerSppg:
          type: object
          properties:
            sppgId:
              type: string
            sppgName:
              type: string
            address:
              type: string
        totalReports:
          type: integer
      required:
        - schoolName
        - partnerSppg
        - totalReports
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
      required:
        - message